refactord out of index.js
